InstUtils, MelodyUtils, MidiUtils refactor; warning fixes; 24 melody blocks
