+++ midimasterpiece/src/main/java/org/vibehistorian/vibecomposer/Helpers/BooleanRenderer.java
@@ -5,14 +5,12 @@ import java.awt.Component;
 import javax.swing.JCheckBox;
 import javax.swing.JPanel;
 import javax.swing.JTable;
-import javax.swing.border.Border;
-import javax.swing.border.EmptyBorder;
 import javax.swing.table.TableCellRenderer;
 
 public class BooleanRenderer extends JPanel implements TableCellRenderer {
 
 	private static final long serialVersionUID = 3986897083726263464L;
-	private static final Border noFocusBorder = new EmptyBorder(1, 1, 1, 1);
+	//private static final Border noFocusBorder = new EmptyBorder(1, 1, 1, 1);
 
 	private JCheckBox checkbox;
