+++ midimasterpiece/src/main/java/org/vibehistorian/vibecomposer/Parts/ChordPart.java
@@ -3,7 +3,7 @@ package org.vibehistorian.vibecomposer.Parts;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
 
-import org.vibehistorian.vibecomposer.MidiUtils.POOL;
+import org.vibehistorian.vibecomposer.InstUtils;
 import org.vibehistorian.vibecomposer.Enums.PatternJoinMode;
 
 @XmlRootElement(name = "chordPart")
@@ -17,7 +17,7 @@ public class ChordPart extends InstPart {
 	private PatternJoinMode patternJoinMode = PatternJoinMode.NOJOIN;
 	private int noteLengthMultiplier = 100;
 
-	private POOL instPool = POOL.PLUCK;
+	private InstUtils.POOL instPool = InstUtils.POOL.PLUCK;
 
 	public ChordPart() {
 
@@ -49,11 +49,11 @@ public class ChordPart extends InstPart {
 	}
 
 
-	public POOL getInstPool() {
+	public InstUtils.POOL getInstPool() {
 		return instPool;
 	}
 
-	public void setInstPool(POOL instPool) {
+	public void setInstPool(InstUtils.POOL instPool) {
 		this.instPool = instPool;
 	}
